+++ router.js
@@ -189,11 +189,11 @@ router.prototype._handle = function(req, res) {
       }
     }
     if (r.path.includes(':')) {
-      parsed = utils.params(r.path, req.url);
+      parsed = utils.params(r.path, req.path);
       if (parsed.params) req.params = parsed.params;
     }
-    if (!utils.wildcard((parsed.path ? parsed.path : r.path), (r.path.endsWith('/') ? req.url : (req.path.substr(1).endsWith('/') ? req.path.slice(0, -1) : req.path)))) continue;
-    /// if (r.method === req.method || r.method === 'GET' && req.method === 'HEAD') {
+    if (!utils.wildcard((parsed.path ? parsed.path : r.path), (r.path.endsWith('/') ? req.path : (req.path.substr(1).endsWith('/') ? req.path.slice(0, -1) : req.path)))) continue;
+    console.log(req.path);
     if (r.method === req.method) {
       r.route(req, res);
       return true;